feat: move filterByType in the sidebar
