+++ packages/web/src/components.d.ts
@@ -10,13 +10,16 @@ declare module 'vue' {
     ContactButton: typeof import('./components/contact/ContactButton.vue')['default']
     ContactHelp: typeof import('./components/contact/ContactHelp.vue')['default']
     ContactMail: typeof import('./components/contact/ContactMail.vue')['default']
+    copy: typeof import('./components/program/list/ProgramFilters copy.vue')['default']
     DsfrButton: typeof import('@gouvminint/vue-dsfr')['DsfrButton']
     DsfrSelect: typeof import('@gouvminint/vue-dsfr')['DsfrSelect']
     ProgramAccordion: typeof import('./components/program/detail/ProgramAccordion.vue')['default']
     ProgramCard: typeof import('./components/program/list/ProgramCard.vue')['default']
     ProgramDetail: typeof import('./components/program/detail/ProgramDetail.vue')['default']
     ProgramEligibility: typeof import('./components/program/detail/ProgramEligibility.vue')['default']
+    ProgramFilterByType: typeof import('./components/program/list/ProgramFilterByType.vue')['default']
     ProgramFilters: typeof import('./components/program/list/ProgramFilters.vue')['default']
+    ProgramFiltersByType: typeof import('./components/program/list/ProgramFiltersByType.vue')['default']
     ProgramForm: typeof import('./components/program/detail/ProgramForm.vue')['default']
     ProgramList: typeof import('./components/program/list/ProgramList.vue')['default']
     ProgramListHeaderResult: typeof import('./components/program/list/ProgramListHeaderResult.vue')['default']
