+++ packages/web/src/components/program/list/ProgramFilters.vue
@@ -1,11 +1,5 @@
 <template>
   <div class="fr-grid-row fr-grid-row--gutters fr-mb-4v">
-    <div class="fr-col-12 fr-col-sm-6">
-      <DsfrSelect
-        v-model="programFilters.programAidTypeSelected"
-        :options="programAidTypeOptions"
-      />
-    </div>
     <div
       v-if="programStore.hasObjectiveTypeFilter()"
       class="fr-col-12 fr-col-sm-6"
@@ -20,40 +14,13 @@
 
 <script setup lang="ts">
 import { useProgramStore } from '@/stores/program'
-import { PublicodeObjective, ProgramAidType, type programFiltersType } from '@/types'
+import { PublicodeObjective, type programFiltersType } from '@/types'
 import { DsfrSelect } from '@gouvminint/vue-dsfr'
 import type { DsfrSelectProps } from '@gouvminint/vue-dsfr/types/components/DsfrSelect/DsfrSelect.types'
 
 const programFilters: programFiltersType = useProgramStore().programFilters
 
 const programStore = useProgramStore()
-
-const programAidTypeOptions: DsfrSelectProps['options'] = [
-  {
-    text: "Filtrer par nature de l'aide",
-    value: ''
-  },
-  {
-    text: 'Accompagnement',
-    value: ProgramAidType.acc
-  },
-  {
-    text: 'Financement',
-    value: ProgramAidType.fund
-  },
-  {
-    text: 'Prêt',
-    value: ProgramAidType.loan
-  },
-  {
-    text: 'Avantage fiscal',
-    value: ProgramAidType.tax
-  },
-  {
-    text: 'Formation',
-    value: ProgramAidType.train
-  }
-]
 
 const objectifTypeOptions: DsfrSelectProps['options'] = [
   {
