+++ packages/web/src/components/questionnaire/TeeQuestionnaireResult.vue
@@ -16,6 +16,7 @@
           style="height: 100%"
         >
           <TrackSidebar />
+          <ProgramFilterByType />
         </div>
 
         <div
